Move macros to dedicated folder
Avoids dealing with custom module resolution
or accidentally publishing the file
