+++ packages/material-ui-utils/test/.eslintrc.js
@@ -4,9 +4,5 @@ module.exports = {
     'import/newline-after-import': 'off',
     // for testing purposes
     'no-useless-concat': 'off',
-    // Need to use explicit /src folder.
-    // Can't use same module resolution as root config.
-    // https://github.com/babel-utils/babel-plugin-tester/issues/74
-    'no-restricted-imports': 'off',
   },
 };
