+++ packages/material-ui-utils/test/MuiError.macro.test.js
@@ -11,7 +11,7 @@ pluginTester({
       title: 'can throw on missing error codes',
       error: /unknown: Missing error code for message 'missing'. Did you forget to run `yarn extract-errors` first?/,
       code: `
-        import MuiError from '@material-ui/utils/src/MuiError.macro';
+        import MuiError from '@material-ui/utils/macros/MuiError.macro';
 
         throw new MuiError('missing');`,
       pluginOptions: {
