+++ packages/material-ui-utils/test/__fixtures__/no-error-code/code.js
@@ -1,3 +1,3 @@
-import MuiError from '@material-ui/utils/src/MuiError.macro';
+import MuiError from '@material-ui/utils/macros/MuiError.macro';
 
 throw new MuiError('Material-UI: Expected valid input target. \n' + 'Did you use inputComponent');
